requirements.txt
matplotlib==3.8.0
numpy==1.26.0
scipy==1.13.0
requests==2.31.0
python-dotenv==1.0.1

